// src/app/pages/approval/models/IOrderDetail.ts
export interface IOrderDetail {
	Descripcion: string;
	sede: string;
	CompaniaSocio: string;
	NumeroOrden: string;
	Proveedor: number;
	NombreCompleto: string;
	FormadePago: string;
	FechaPreparacion: string;
	FechaAprobacion: string | null;
	ImpresionNumero: number;
	Clasificacion: string;
	UnidadNegocio: string;
	Observaciones: string;
	RacionFlag: string;
	LocalForeignFlag: string;
	AlmacenCodigo: string;
	TipodeCambio: number;
	MonedaCodigo: string;
	MontoTotal: number;
	MontoOtros: number;
	MontoFlete: number | null;
	Estado: string;
	UltimaFechaModif: string;
	UltimoUsuario: string;
	MontoPendientedePago: number;
	AlmacenCodigoIngreso: string;
	PlazoEntrega: number;
	ObservacionesEntrega: string | null;
	NumeroInterno: string | null;
	RazonRechazo: string | null;
	RequisicionNumero: string;
	NumeroContrato: string | null;
	FechaPrometida: string;
	FechaEnvioProveedor: string | null;
	Sucursal: string;
	MontoBruto: number;
	MontoNoAfecto: number;
	DescripcionLocal: string;
	NivelAprobacion: string;

	desformapag: string;
	TipoCompromiso: string;
	NumeroCompromiso: string;
	CodigoInterno: string | null;
	CompaniaCodigo: string;
	MonedaDocumento: string;
	MonedaPago: string;
	TipoPago: string;
	FechaEntrega: string;
	FechaDocumento: string;
	DiasparaPago: number;
	TipodeServicio: string;
	TipodeServiciocxp: string;
	TipodeCotizacion: string;
	MontoAfecto: number;
	MontoCompromisoOriginal: number;
	MontoCompromiso: number;
	MontoImpuesto: number;
	MontoGastado: number;
	ProvisionTipoFlag: string | null;
	PreparadaPor: number;
	AprobadaPor: number;
	EstadoCompromiso: string;
	ContraPrestacionFlag: string;
	CPPendienteFlag: string;
	Periodicidad: string | null;
	PeriodoAplicacion: string | null;
	PrecioUnitario: number;
	Comentarios: string | null;
	Busqueda: string;
	MontoIGV: number;
}